==> code/step_1_plot_raw_data.rmd <==
---
title: "Plot Raw Data"
author: "Jonathan Salas"
date: "Last Updated `r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

# Start with a clean slate

```{r echo = F, include=FALSE}
# Remove all starting code
rm(list = ls(all.names = TRUE)) 
# Detach Packages 
try(invisible(lapply(paste0('package:', names(sessionInfo()$otherPkgs)), detach, character.only=TRUE, unload=TRUE)), silent = T) # Detach Packages
```

# Load Packages 

```{r setup, include=FALSE}
library(eiEAM)
library(rgl)
library(knitr)
library(plotly)
library(dplyr)
opts_knit$set(root.dir = "~/R_Projects/easieR_CV/")
```

# What subject and timepoint are we working with?

```{r}
id <- "sbrt_03" # all lowercase, study_id format e.g., "sbrt_03"
timepoint <- "t1" # "t1" vs "t2" t1 = baseline; t2 = terminal
```

# Import the raw data

```{r}
# From step 1: 
load(paste0("data/unaltered_data/",id,"_",timepoint,"_raw.RData"))
data_1 <- raw.data.1
load(paste0("data/unaltered_data/",id,"_",timepoint,"_raw_volt_connex.RData"))
```

# Plot it:

```{r}
# Create mesh3d object with cleaning instead of cgalMesh
voltage_mesh <- tmesh3d(
  vertices = t(as.matrix(carto_lv_sr[, 1:3])),
  indices = t(as.matrix(carto_lv_sr_connex[, 1:3])),
  homogeneous = FALSE
)

# Generate wireframe coordinates using Rvcg-compatible function
voltage_wireframe <- create_wireframe_mesh(voltage_mesh)

# Create mesh3d object with cleaning instead of cgalMesh
apaced_mesh <- tmesh3d(
  vertices = t(as.matrix(data_1 |> filter(group == "carto_mesh", pacing_dir == "apaced") |> select(surf_x:surf_z))),
  indices = t(as.matrix(carto_connex |> filter(pacing_dir == "apaced") |> select(v1:v3))),
  homogeneous = FALSE
)

# Generate wireframe coordinates using Rvcg-compatible function
apaced_wireframe <- create_wireframe_mesh(apaced_mesh)

# Create mesh3d object with cleaning instead of cgalMesh
rvpaced_mesh <- tmesh3d(
  vertices = t(as.matrix(data_1 |> filter(group == "carto_mesh", pacing_dir == "rvpaced") |> select(surf_x:surf_z))),
  indices = t(as.matrix(carto_connex |> filter(pacing_dir == "rvpaced") |> select(v1:v3))),
  homogeneous = FALSE
)

# Generate wireframe coordinates using Rvcg-compatible function
rvpaced_wireframe <- create_wireframe_mesh(rvpaced_mesh)
```

```{r echo=F}
plot_ly() |>
  add_trace(
    name = "Voltage Wireframe",
    data = voltage_wireframe,
    x = ~ x,
    y = ~ y,
    z = ~ z,
    type = "scatter3d",
    mode = "lines",
    line = list(color = "black", width = 3),
    showlegend = TRUE,
    hoverinfo = "none"
  ) |>
  add_trace(
    name = "CARTO Mesh Vertices - Voltage",
    type = "scatter3d",
    mode = "markers",
    data = carto_lv_sr,
    x = ~ surf_x,
    y = ~ surf_y,
    z = ~ surf_z,
    showlegend = T
  ) |>
  add_trace(
    name = "Voltage",
    type = "scatter3d",
    mode = "markers",
    data = data_1 |> filter(group == "lv_volt"),
    x = ~ x,
    y = ~ y,
    z = ~ z,
    showlegend = TRUE
  ) |>
  add_trace(
    name = "Atrial Paced Wireframe",
    data = apaced_wireframe,
    x = ~ x,
    y = ~ y,
    z = ~ z,
    type = "scatter3d",
    mode = "lines",
    line = list(color = "black", width = 3),
    showlegend = TRUE,
    hoverinfo = "none"
  ) |>
  add_trace(
    name = "CARTO Mesh Vertices - Atrial Paced",
    type = "scatter3d",
    mode = "markers",
    data = data_1 |> filter(group == "carto_mesh" &
                              pacing_dir == "apaced"),
    x = ~ surf_x,
    y = ~ surf_y,
    z = ~ surf_z,
    showlegend = T
  ) |>
  add_trace(
    name = "LAT - Atrial Paced",
    type = "scatter3d",
    mode = "markers",
    data = data_1 |> filter(group == "lat_coord" &
                              pacing_dir == "apaced"),
    x = ~ x,
    y = ~ y,
    z = ~ z,
    showlegend = T
  ) |>
  add_trace(
    name = "Scar fit to Atrial Paced Mesh",
    type = "scatter3d",
    mode = "markers",
    data = data_1 |> filter(group == "mri_scar" &
                              pacing_dir == "apaced"),
    x = ~ x,
    y = ~ y,
    z = ~ z
  ) |>
  add_trace(
    name = "RV Paced Wireframe",
    data = rvpaced_wireframe,
    x = ~ x,
    y = ~ y,
    z = ~ z,
    type = "scatter3d",
    mode = "lines",
    line = list(color = "black", width = 3),
    showlegend = TRUE,
    hoverinfo = "none"
  ) |>
  add_trace(
    name = "CARTO Mesh Vertices - RV Paced",
    type = "scatter3d",
    mode = "markers",
    data = data_1 |> filter(group == "carto_mesh" &
                              pacing_dir == "rvpaced"),
    x = ~ surf_x,
    y = ~ surf_y,
    z = ~ surf_z,
    showlegend = T
  ) |>
  add_trace(
    name = "LAT - RV Paced",
    type = "scatter3d",
    mode = "markers",
    data = data_1 |> filter(group == "lat_coord" &
                              pacing_dir == "rvpaced"),
    x = ~ x,
    y = ~ y,
    z = ~ z,
    showlegend = T
  ) |>
  add_trace(
    name = "Scar fit to RV Paced Mesh",
    type = "scatter3d",
    mode = "markers",
    data = data_1 |> filter(group == "mri_scar" &
                              pacing_dir == "rvpaced"),
    x = ~ x,
    y = ~ y,
    z = ~ z
  ) |>
  layout(
    showlegend = TRUE,
    scene = list(camera = list(
      eye = list(x = -1.5, y = 0.2, z = 1.5),
      # orbital angle
      up  = list(x = 0, y = 1, z = 0)      # which direction is "up"
    )),
    title = list(
      text = paste0(
        "ID:",
        id,
        "<br>",
        "Timepoint: ",
        timepoint,
        "<br>",
        "Intervention: ",
        unique(data_1$intervention)
      )
    )
  )
```


# Version and Package Details

```{r echo=F}
paste(version$version.string, version$nickname)
paste0(
  "RStudio Version ",
  rstudioapi::versionInfo()$version,
  " ",
  rstudioapi::versionInfo()$release_name
) # RStudio Version
subset(data.frame(sessioninfo::package_info()),
       attached == TRUE,
       c(package, loadedversion)) # Loaded Packages
```

# When were these files last rewritten?

```{r}
date()
```
